Some fixes, improved readme
Readme should explain better what to do with this (test it before
shipping).

Separated the script in two parts, the immutable part (that will be
changed when adding better control and multiple input/output filenames)
and edit_html which contains the main function that edits the html
before writing it into pdf. This function is the one that should be
edited mostly to account for the changes that should be applied to the
DOM of the html file before its converted to pdf.

Added an empty init file since this is what you do for packages
apparently? Maybe?