Bug fixes, changes to structure
Small changes to structure to reflect better what can be edited and
where the aux functions are.

Some fun bugfixes (like having to check that body exists in an html...
it should but w/e, or valid filenames excluding letters with accents
(not sure if its very legal and very cool but my usecases will probably
include some, so I do have to accept them)

The important stuff:
Okay, so wkhtmltopdf wasn't working with a big page that had relative
imports and pretty complex structure/images/stylesheets/scripts, etc.

- Apparently, this was related to the page having relative imports
instead of absolute ones. Created a small function to change them but it
did not help.

- Tried some flags for pdfkit, etc. Still not working.

- Experimented with wkhtmltopdf in the cli. It gives better error info
that way. Apparently the error was because a js for a video loader
embedded something in the page dynamically (an iframe?) with some
attribute (I think x-origin)=file:///youtubei/v1/log_event={big json
with some API uuid}.
It put file:// since the script was using window.location.origin to get
the path.
So in the end wkhtml thought it was something it had to load so it tried
and failed hard, and that was what was borking everything.

- Removed js from loading when doing the conversion, it shouldn't be
needed.

- Everything seems to work so far, some warnings from the stylesheets
using url(non found import that only exists in the ioc server) but
nothing serious. Removed absolute imports.

- Need to check if this all works fine with docker